Move withMods scope to main score model
