+++ app/Models/Score/Best/Model.php
@@ -5,7 +5,6 @@
 
 namespace App\Models\Score\Best;
 
-use App\Libraries\ModsHelper;
 use App\Libraries\ReplayFile;
 use App\Models\Beatmap;
 use App\Models\ReplayViewCount;
@@ -207,22 +206,6 @@ abstract class Model extends BaseModel
     public function scopeVisibleUsers($query)
     {
         return $query->where(['hidden' => false]);
-    }
-
-    public function scopeWithMods($query, $modsArray)
-    {
-        return $query->where(function ($q) use ($modsArray) {
-            $bitset = ModsHelper::toBitset($modsArray);
-            $preferenceMask = ~ModsHelper::PREFERENCE_MODS_BITSET;
-
-            if (in_array('NM', $modsArray, true)) {
-                $q->orWhereRaw('enabled_mods & ? = 0', [$preferenceMask]);
-            }
-
-            if ($bitset > 0) {
-                $q->orWhereRaw('enabled_mods & ? = ?', [$preferenceMask | $bitset, $bitset]);
-            }
-        });
     }
 
     public function scopeWithType($query, $type, $options)
